requirements.txt
Flask==3.0.2
openai==0.28.1
python-dotenv==1.0.1
Werkzeug==3.0.1
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.27 